requirements.txt
beautifulsoup4
requests
lxml
pyyaml
ipwhois
geopip2